Compromise between Ben's changes to my code and the thing I was thinking when I wrote it.
I initially had a useless variable named Health in the constructor, so that it was obvious that the number in the super() constructor referred to health (I have a very strong dislike for magic numbers). This is inefficient, so Ben correctly changed it. I revised it further by simply adding a comment about the significance of 80. I might be a bit pedantic.